eembed_log to favor 64 bit
No real advantage in replacing "long" with "long long" in order to
get 64 bit on 32bit and smaller platforms.

This will mean that printing an int on a smaller platform may be more
expensive than the native "long" time, but not _much_ more expensive
for most platforms.

eembed-arduino.c: align types to log, but cast-narrow "int64_t" to
"long" for printing, with small potential for issues if anyone is
actually wanting to print large int64_t values from the arduino
firmware, which Serial.print does not support.

+++ src/eembed-arduino.cpp
@@ -39,10 +39,10 @@ void serial_log_print_s(struct eembed_log *log, const char *s)
 	Serial.print(s);
 }
 
-void serial_log_print_ul(struct eembed_log *log, unsigned long ul)
+void serial_log_print_ul(struct eembed_log *log, uint64_t ul)
 {
 	(void)log;
-	Serial.print(ul);
+	Serial.print((unsigned long)ul);
 }
 
 void serial_log_print_z(struct eembed_log *log, size_t z)
@@ -51,16 +51,16 @@ void serial_log_print_z(struct eembed_log *log, size_t z)
 	Serial.print(z);
 }
 
-void serial_log_print_l(struct eembed_log *log, long l)
+void serial_log_print_l(struct eembed_log *log, int64_t l)
 {
 	(void)log;
-	Serial.print(l);
+	Serial.print((long)l);
 }
 
-void serial_log_print_f(struct eembed_log *log, double f)
+void serial_log_print_f(struct eembed_log *log, long double f)
 {
 	(void)log;
-	Serial.print(f);
+	Serial.print((double)f);
 }
 
 void serial_log_print_vp(struct eembed_log *log, const void *v)
